Make residue traits public
