+++ src/uint/modular/constant_mod/const_inv.rs
@@ -3,7 +3,7 @@ use core::marker::PhantomData;
 use subtle::{Choice, CtOption};
 
 use crate::{
-    modular::inv::{inv_montgomery_form, InvResidue},
+    modular::{inv::inv_montgomery_form, InvResidue},
     Word,
 };
