+++ src/uint/modular/mul.rs
@@ -2,18 +2,6 @@ use crate::{Limb, UInt};
 
 use super::reduction::montgomery_reduction;
 
-pub trait MulResidue
-where
-    Self: Sized,
-{
-    /// Computes the (reduced) product of two residues.
-    fn mul(&self, rhs: &Self) -> Self;
-
-    fn square(&self) -> Self {
-        self.mul(self)
-    }
-}
-
 pub(crate) const fn mul_montgomery_form<const LIMBS: usize>(
     a: &UInt<LIMBS>,
     b: &UInt<LIMBS>,
